fix(babel): Alternative to for/of and make es2015 preset prod dep
The for/of transpilation in this package is not compatible with React Native. However when writing

for/of loops in React Native project, their transpilation works just fine. Will need to look into

this more. es2015 preset needed by React Native. Instead of forcing user of package to install

babel-preset-es2015, I am moving it from devDependencies to dependencies(prod that is).

closes #5 #6